+++ crates/metasearch-engine/src/postgres_engine.rs
@@ -108,7 +108,7 @@ impl SearchEngine for PostgresEngine {
         {
             use tokio_postgres::{NoTls, Row};
 
-            let offset = (query.page.saturating_sub(1)) * self.limit;
+            let offset = (query.page.saturating_sub(1)) * (self.limit as u32);
             let wildcard = format!("%{}%", query.query.replace(' ', "%"));
 
             let conn_str = format!(
